fix(ci): python coverage shard line limit — split oversized PR monitor tests
